refactor utility codes; refactor schemas
