+++ app/main.py
@@ -1,132 +1,22 @@
 import httpx
-import time
-import calendar
 import re
 
 from fastapi import FastAPI, HTTPException, status
 from pydantic import BaseModel
 from typing import Dict, List, Optional
-from datetime import date
+from .utils import constant, utility_functions
+from .schemas import entrypoint_schema
+from .schemas import yearly_schema
+from .schemas import monthly_schema
+from .schemas import daily_schema
 
-def is_date_greater_than(date1, date2):
-    formatted_date1 = time.strptime(date1, "%Y-%m-%d")
-    formatted_date2 = time.strptime(date2, "%Y-%m-%d")
 
-    return formatted_date1 >= formatted_date2
-
-def is_date_lower_than(date1, date2):
-    formatted_date1 = time.strptime(date1, "%Y-%m-%d")
-    formatted_date2 = time.strptime(date2, "%Y-%m-%d")
-
-    return formatted_date1 <= formatted_date2
-
-def is_year_month_greater_than(year_month1, year_month2):
-    formatted_date1 = time.strptime(year_month1+"-01", "%Y-%m-%d")
-    formatted_date2 = time.strptime(year_month2+"-01", "%Y-%m-%d")
-
-    return formatted_date1 >= formatted_date2
-
-def is_year_month_lower_than(year_month1, year_month2):
-    formatted_date1 = time.strptime(year_month1+"-01", "%Y-%m-%d")
-    formatted_date2 = time.strptime(year_month2+"-01", "%Y-%m-%d")
-
-    return formatted_date1 <= formatted_date2
-
-def string_formatted_day(year, month):
-    day = calendar.monthrange(int(year), int(month))
-
-    formatted_day = "-" + str(day[1])
-
-    return formatted_day
 
 app = FastAPI()
 
 covid_data_JSON = {}
 with httpx.Client() as client:
     covid_data_JSON = client.get("https://data.covid19.go.id/public/api/update.json").json()
-
-EARLIEST_YEAR_CASE = "2020"
-EARLIEST_MONTH_CASE = "03"
-EARLIEST_DAY_CASE = "02"
-
-EARLIEST_YEAR_MONTH_CASE = EARLIEST_YEAR_CASE + "." + EARLIEST_MONTH_CASE
-EARLIEST_DATE_CASE = EARLIEST_YEAR_MONTH_CASE + "." + EARLIEST_DAY_CASE
-
-todays_date = date.today()
-
-LATEST_YEAR_CASE = str(todays_date.year)
-LATEST_MONTH_CASE = (str(int(todays_date.month))) if (int(todays_date.month) >= 10) else ("0" + str(int(todays_date.month)))
-LATEST_DAY_CASE = (str(int(todays_date.day))) if (int(todays_date.day) >= 10) else ("0" + str(int(todays_date.day)))
-
-LATEST_YEAR_MONTH_CASE = LATEST_YEAR_CASE + "." + LATEST_MONTH_CASE
-LATEST_DATE_CASE = LATEST_YEAR_MONTH_CASE + "." + LATEST_DAY_CASE
-
-
-class EntrypointResponseData(BaseModel):
-    total_positive: int
-    total_recovered: int
-    total_deaths: int
-    total_active: int
-    new_positive: int
-    new_recovered: int
-    new_deaths: int
-    new_active: int
-
-class EntrypointResponse(BaseModel):
-    ok: bool
-    data: EntrypointResponseData
-    message: str
-
-class YearlyResponseData(BaseModel):
-    year: str
-    positive: int
-    recovered: int
-    deaths: int
-    active: int
-
-class YearlyResponseMulti(BaseModel):
-    ok: bool
-    data: List[YearlyResponseData]
-    message: str
-
-class YearlyResponseSingle(BaseModel):
-    ok: bool
-    data: YearlyResponseData
-    message: str
-
-class MonthlyResponseData(BaseModel):
-    month: str
-    positive: int
-    recovered: int
-    deaths: int
-    active: int
-
-class MonthlyResponseMulti(BaseModel):
-    ok: bool
-    data: List[MonthlyResponseData]
-    message: str
-
-class MonthlyResponseSingle(BaseModel):
-    ok: bool
-    data: MonthlyResponseData
-    message: str
-
-class DailyResponseData(BaseModel):
-    date: str
-    positive: int
-    recovered: int
-    deaths: int
-    active: int
-
-class DailyResponseMulti(BaseModel):
-    ok: bool
-    data: List[DailyResponseData]
-    message: str
-
-class DailyResponseSingle(BaseModel):
-    ok: bool
-    data: DailyResponseData
-    message: str
 
 def create_single_REST_response():
     return {
@@ -215,7 +105,7 @@ def find_data_by_dates(since, upto):
     result = []
 
     for daily_data in httpx_response_JSON["update"]["harian"]:
-        if((is_date_greater_than(daily_data["key_as_string"][0:10], since)) and (is_date_lower_than(daily_data["key_as_string"][0:10], upto))):
+        if((utility_functions.is_date_greater_than(daily_data["key_as_string"][0:10], since)) and (utility_functions.is_date_lower_than(daily_data["key_as_string"][0:10], upto))):
             result.append({
                 "date": daily_data["key_as_string"][0:10],
                 "positive": daily_data["jumlah_positif"]["value"],
@@ -228,8 +118,8 @@ def find_data_by_dates(since, upto):
 
 def regex_date_query_generator(type):
     regex_string = "("
-    for year in range(int(EARLIEST_YEAR_CASE), int(LATEST_YEAR_CASE) + 1, 1):
-        if year == int(LATEST_YEAR_CASE):
+    for year in range(int(constant.EARLIEST_YEAR_CASE), int(constant.LATEST_YEAR_CASE) + 1, 1):
+        if year == int(constant.LATEST_YEAR_CASE):
             regex_string += str(year) + ").("
         else:
             regex_string += str(year) + "|"
@@ -259,7 +149,7 @@ def regex_date_query_generator(type):
 
 
 
-@app.get("/", response_model=EntrypointResponse)
+@app.get("/", response_model=entrypoint_schema.EntrypointResponse)
 async def index():
     httpx_response_JSON = httpx.get("https://data.covid19.go.id/public/api/update.json").json()
     response_body = create_single_REST_response()
@@ -276,14 +166,14 @@ async def index():
     }
     return response_body
 
-@app.get("/yearly", response_model=YearlyResponseMulti)
-async def yearly_multi(since: Optional[int] = int(EARLIEST_YEAR_CASE), upto: Optional[int] = int(LATEST_YEAR_CASE)):
+@app.get("/yearly", response_model=yearly_schema.YearlyResponseMulti)
+async def yearly_multi(since: Optional[int] = int(constant.EARLIEST_YEAR_CASE), upto: Optional[int] = int(constant.LATEST_YEAR_CASE)):
     response_body = create_multi_REST_response()
 
-    if(since < int(EARLIEST_YEAR_CASE) or upto > int(LATEST_YEAR_CASE)):
+    if(since < int(constant.EARLIEST_YEAR_CASE) or upto > int(constant.LATEST_YEAR_CASE)):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Query parameters error! SINCE must be higher or equal than " + EARLIEST_YEAR_CASE +" and UPTO can't be higher than " + LATEST_YEAR_CASE +"!"
+            detail="Query parameters error! SINCE must be higher or equal than " + constant.EARLIEST_YEAR_CASE +" and UPTO can't be higher than " + constant.LATEST_YEAR_CASE +"!"
         )
 
     if(since > upto):
@@ -296,18 +186,18 @@ async def yearly_multi(since: Optional[int] = int(EARLIEST_YEAR_CASE), upto: Opt
 
     return response_body
 
-@app.get("/yearly/{year}", response_model=YearlyResponseSingle)
+@app.get("/yearly/{year}", response_model=yearly_schema.YearlyResponseSingle)
 async def yearly_single(year: str):
     if(not year.isnumeric()):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Value error! the provided YEAR must be an integer! with a value between " + EARLIEST_YEAR_CASE +" and " + LATEST_YEAR_CASE +"!"
+            detail="Value error! the provided YEAR must be an integer! with a value between " + constant.EARLIEST_YEAR_CASE +" and " + constant.LATEST_YEAR_CASE +"!"
         )
 
-    if(int(year) < int(EARLIEST_YEAR_CASE) or int(year) > int(LATEST_YEAR_CASE)):
+    if(int(year) < int(constant.EARLIEST_YEAR_CASE) or int(year) > int(constant.LATEST_YEAR_CASE)):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Value error! the provided YEAR must be higher or equal than " + EARLIEST_YEAR_CASE +" and can't be higher than " + LATEST_YEAR_CASE +"!"
+            detail="Value error! the provided YEAR must be higher or equal than " + constant.EARLIEST_YEAR_CASE +" and can't be higher than " + constant.LATEST_YEAR_CASE +"!"
         )
 
     response_body = create_single_REST_response()
@@ -315,13 +205,13 @@ async def yearly_single(year: str):
 
     return response_body
 
-@app.get("/monthly", response_model=MonthlyResponseMulti)
-async def monthly_multi(since: Optional[str] = EARLIEST_YEAR_MONTH_CASE, upto: Optional[str] = LATEST_YEAR_MONTH_CASE):
+@app.get("/monthly", response_model=monthly_schema.MonthlyResponseMulti)
+async def monthly_multi(since: Optional[str] = constant.EARLIEST_YEAR_MONTH_CASE, upto: Optional[str] = constant.LATEST_YEAR_MONTH_CASE):
 
     if (not re.search(regex_date_query_generator("MONTH"), since) or not re.search(regex_date_query_generator("MONTH"), upto)):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Query parameter error! SINCE and UPTO must be a valid date in the format of 'YYYY.MM'. Valid values for YYYY are the value of integers between " + EARLIEST_YEAR_CASE + " and " + LATEST_YEAR_CASE)
+            detail="Query parameter error! SINCE and UPTO must be a valid date in the format of 'YYYY.MM'. Valid values for YYYY are the value of integers between " + constant.EARLIEST_YEAR_CASE + " and " + constant.LATEST_YEAR_CASE)
 
 
     response_body = create_multi_REST_response()
@@ -345,12 +235,12 @@ async def monthly_multi(since: Optional[str] = EARLIEST_YEAR_MONTH_CASE, upto: O
 
     return response_body
     
-@app.get("/monthly/{year}", response_model=MonthlyResponseMulti)
+@app.get("/monthly/{year}", response_model=monthly_schema.MonthlyResponseMulti)
 async def monthly_year_multi(year: str, since: Optional[str] = None, upto: Optional[str] = None):
-    if((not year.isnumeric()) or (int(year) < int(EARLIEST_YEAR_CASE)) or (int(year) > int(LATEST_YEAR_CASE))):
+    if((not year.isnumeric()) or (int(year) < int(constant.EARLIEST_YEAR_CASE)) or (int(year) > int(constant.LATEST_YEAR_CASE))):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail= "the provided YEAR must be an integer value between " + EARLIEST_YEAR_CASE + " and " + LATEST_YEAR_CASE + "!"
+            detail= "the provided YEAR must be an integer value between " + constant.EARLIEST_YEAR_CASE + " and " + constant.LATEST_YEAR_CASE + "!"
             )
 
     response_body = create_multi_REST_response()
@@ -358,8 +248,8 @@ async def monthly_year_multi(year: str, since: Optional[str] = None, upto: Optio
     since_year_month = ""
     if since:
         since_year_month = (since.split("."))
-    elif (year == EARLIEST_YEAR_CASE):
-        since_year_month = ((EARLIEST_YEAR_MONTH_CASE).split("."))
+    elif (year == constant.EARLIEST_YEAR_CASE):
+        since_year_month = ((constant.EARLIEST_YEAR_MONTH_CASE).split("."))
     else:
         since_year_month = ((year + ".01").split("."))
     
@@ -367,8 +257,8 @@ async def monthly_year_multi(year: str, since: Optional[str] = None, upto: Optio
     upto_year_month = ""
     if upto:
         upto_year_month = (upto.split("."))
-    elif (year == LATEST_YEAR_CASE):
-        upto_year_month = ((LATEST_YEAR_MONTH_CASE).split("."))
+    elif (year == constant.LATEST_YEAR_CASE):
+        upto_year_month = ((constant.LATEST_YEAR_MONTH_CASE).split("."))
     else:
         upto_year_month = ((year + ".12").split("."))
     
@@ -395,7 +285,7 @@ async def monthly_year_multi(year: str, since: Optional[str] = None, upto: Optio
     
     return response_body
 
-@app.get("/monthly/{year}/{month}", response_model=MonthlyResponseSingle)
+@app.get("/monthly/{year}/{month}", response_model=monthly_schema.MonthlyResponseSingle)
 async def monthly_year_month_single(year: str, month: str):
     if(not year.isnumeric() or not month.isnumeric()):
         raise HTTPException(
@@ -403,10 +293,10 @@ async def monthly_year_month_single(year: str, month: str):
             detail="YEAR and MONTH must be an integer!"
         )
 
-    if(int(year) < int(EARLIEST_YEAR_CASE) or int(year) > int(LATEST_YEAR_CASE)):
+    if(int(year) < int(constant.EARLIEST_YEAR_CASE) or int(year) > int(constant.LATEST_YEAR_CASE)):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Invalid YEAR value! YEAR must be an integer with a value between " + EARLIEST_YEAR_CASE + " and " + LATEST_YEAR_CASE + "!"
+            detail="Invalid YEAR value! YEAR must be an integer with a value between " + constant.EARLIEST_YEAR_CASE + " and " + constant.LATEST_YEAR_CASE + "!"
         )
     
     if(int(month) < 1 or int(month) > 12):
@@ -421,12 +311,12 @@ async def monthly_year_month_single(year: str, month: str):
 
     return response_body
 
-@app.get("/daily", response_model=DailyResponseMulti)
-async def daily_multi(since: Optional[str] = EARLIEST_DATE_CASE, upto: Optional[str] = LATEST_DATE_CASE):
+@app.get("/daily", response_model=daily_schema.DailyResponseMulti)
+async def daily_multi(since: Optional[str] = constant.EARLIEST_DATE_CASE, upto: Optional[str] = constant.LATEST_DATE_CASE):
     if (not re.search(regex_date_query_generator("DATE"), since) or not re.search(regex_date_query_generator("DATE"), upto)):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Query parameter error! SINCE and UPTO must be a valid date in the format of 'YYYY.MM.DD'. Valid values for YYYY are the value of integers between " + EARLIEST_YEAR_CASE + " and " + LATEST_YEAR_CASE)
+            detail="Query parameter error! SINCE and UPTO must be a valid date in the format of 'YYYY.MM.DD'. Valid values for YYYY are the value of integers between " + constant.EARLIEST_YEAR_CASE + " and " + constant.LATEST_YEAR_CASE)
 
     response_body = create_multi_REST_response()
 
@@ -437,12 +327,12 @@ async def daily_multi(since: Optional[str] = EARLIEST_DATE_CASE, upto: Optional[
 
     return response_body
 
-@app.get("/daily/{year}", response_model=DailyResponseMulti)
+@app.get("/daily/{year}", response_model=daily_schema.DailyResponseMulti)
 async def daily_multi_year(year: str,since: Optional[str] = None, upto: Optional[str] = None):
-    if((not year.isnumeric()) or (int(year) < int(EARLIEST_YEAR_CASE)) or (int(year) > int(LATEST_YEAR_CASE))):
+    if((not year.isnumeric()) or (int(year) < int(constant.EARLIEST_YEAR_CASE)) or (int(year) > int(constant.LATEST_YEAR_CASE))):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail= "the provided YEAR must be an integer value between " + EARLIEST_YEAR_CASE + " and " + LATEST_YEAR_CASE + "!"
+            detail= "the provided YEAR must be an integer value between " + constant.EARLIEST_YEAR_CASE + " and " + constant.LATEST_YEAR_CASE + "!"
         )
 
     response_body = create_multi_REST_response()
@@ -454,15 +344,15 @@ async def daily_multi_year(year: str,since: Optional[str] = None, upto: Optional
         formatted_upto = "-".join((upto.split(".")))
 
     elif (not since and upto):
-        if year == EARLIEST_YEAR_CASE:
-            formatted_since = "-".join((EARLIEST_DATE_CASE.split(".")))
+        if year == constant.EARLIEST_YEAR_CASE:
+            formatted_since = "-".join((constant.EARLIEST_DATE_CASE.split(".")))
         else:
             formatted_since = year + "-01-01"
         formatted_upto = "-".join((upto.split(".")))
     
     elif (since and not upto):
-        if year == LATEST_YEAR_CASE:
-            formatted_upto = "-".join((LATEST_DATE_CASE.split(".")))
+        if year == constant.LATEST_YEAR_CASE:
+            formatted_upto = "-".join((constant.LATEST_DATE_CASE.split(".")))
         else:
             formatted_upto = year + "-12-31"
         formatted_since = "-".join((since.split(".")))
@@ -474,12 +364,12 @@ async def daily_multi_year(year: str,since: Optional[str] = None, upto: Optional
 
     return response_body
 
-@app.get("/daily/{year}/{month}", response_model=DailyResponseMulti)
+@app.get("/daily/{year}/{month}", response_model=daily_schema.DailyResponseMulti)
 async def daily_multi_year_month(year: str, month: str, since: Optional[str] = None, upto: Optional[str] = None):
-    if((not year.isnumeric()) or (int(year) < int(EARLIEST_YEAR_CASE)) or (int(year) > int(LATEST_YEAR_CASE))):
+    if((not year.isnumeric()) or (int(year) < int(constant.EARLIEST_YEAR_CASE)) or (int(year) > int(constant.LATEST_YEAR_CASE))):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail= "the provided YEAR must be an integer value between " + EARLIEST_YEAR_CASE + " and " + LATEST_YEAR_CASE + "!"
+            detail= "the provided YEAR must be an integer value between " + constant.EARLIEST_YEAR_CASE + " and " + constant.LATEST_YEAR_CASE + "!"
         )
     if((not month.isnumeric()) or (int(month) < 1) or (int(month) > 12)):
         raise HTTPException(
@@ -500,21 +390,21 @@ async def daily_multi_year_month(year: str, month: str, since: Optional[str] = N
         formatted_upto = "-".join((upto.split(".")))
     elif(since and not upto):
         formatted_since = "-".join((since.split(".")))
-        formatted_upto = year + "-" + month + string_formatted_day(year, month)
+        formatted_upto = year + "-" + month + utility_functions.string_formatted_day(year, month)
     else:
         formatted_since = year + "-" + month + "-01"
-        formatted_upto = year + "-" + month + string_formatted_day(year, month)
+        formatted_upto = year + "-" + month + utility_functions.string_formatted_day(year, month)
     
     response_body["data"] = find_data_by_dates(formatted_since, formatted_upto)
 
     return response_body
 
-@app.get("/daily/{year}/{month}/{date}", response_model=DailyResponseSingle)
+@app.get("/daily/{year}/{month}/{date}", response_model=daily_schema.DailyResponseSingle)
 async def daily_multi_year_month_date(year: str, month: str, date: str):
-    if((not year.isnumeric()) or (int(year) < int(EARLIEST_YEAR_CASE)) or (int(year) > int(LATEST_YEAR_CASE))):
+    if((not year.isnumeric()) or (int(year) < int(constant.EARLIEST_YEAR_CASE)) or (int(year) > int(constant.LATEST_YEAR_CASE))):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail= "the provided YEAR must be an integer value between " + EARLIEST_YEAR_CASE + " and " + LATEST_YEAR_CASE + "!"
+            detail= "the provided YEAR must be an integer value between " + constant.EARLIEST_YEAR_CASE + " and " + constant.LATEST_YEAR_CASE + "!"
         )
     if((not month.isnumeric()) or (int(month) < 1) or (int(month) > 12)):
         raise HTTPException(
